services/business/users-auth/requirements.txt
aiokafka
aerich
fastapi
tortoise-orm
tortoise-orm[asyncpg]
PyJWT
redis
pytest
python-dotenv
passlib
python-multipart
email-validator
bcrypt
cryptography
uvicorn
prometheus-fastapi-instrumentator>=6.1.0
prometheus-client
hvac
load_dotenv